Notification codes must be greater than zero
The docs say the notification codes can be greater than or equal to 0, but the actual implementation has a check for <= 0. Assuming the implementation is correct, the docs should be updated. I'm not sure if notification code 0 is actually used for anything by default, though.

runtime_lua_nakama.go
```lua
	code := l.CheckInt(4)
	if code <= 0 {
		l.ArgError(4, "expects code to number above 0")
		return 0
	}
```

runtime_go_nakama.go
```go
	if notification.Code <= 0 {
		return errors.New("expects code to number above 0")
	}
```